add DocumentWriter implementation
